rmarkdown::html_vignette:     highlight: "declared.theme"     css: "declared.css"
is commented out because even though it works perfectly with Rmarkdown, pandoc, devtools, somehow it creates a conflict with the pkgdown package builders defaults css settings.

I am sure that this can be fixed relatively easily, most probably the css has to be placed somewhere else, but I cannot research this now.